[DEV-2916] Refactoring
