+++ src/subdomains/core/buy-crypto/buy-crypto.module.ts
@@ -20,6 +20,7 @@ import { PricingModule } from 'src/subdomains/supporting/pricing/pricing.module'
 import { AmlModule } from '../aml/aml.module';
 import { LiquidityManagementModule } from '../liquidity-management/liquidity-management.module';
 import { SellCryptoModule } from '../sell-crypto/sell-crypto.module';
+import { TransactionUtilModule } from '../transaction/transaction-util.module';
 import { BuyCryptoController } from './process/buy-crypto.controller';
 import { BuyCryptoBatch } from './process/entities/buy-crypto-batch.entity';
 import { BuyCryptoFee } from './process/entities/buy-crypto-fees.entity';
@@ -67,6 +68,7 @@ import { SwapService } from './routes/swap/swap.service';
     FiatPayInModule,
     SiftModule,
     FiatOutputModule,
+    TransactionUtilModule,
   ],
   controllers: [BuyCryptoController, BuyController, CryptoRouteController, SwapController],
   providers: [
